% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_cpp_sample.R
\name{popcount}
\alias{popcount}
\title{Count 1's in each element}
\usage{
popcount(xs)
}
\arguments{
\item{xs}{A raw or integer vector to count populations}
}
\value{
The populations of elements in the vector
}
\description{
Count 1's in each element
}
