% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popcount_cpp_integer}
\alias{popcount_cpp_integer}
\title{Count 1's in each integer element}
\usage{
popcount_cpp_integer(xs)
}
\arguments{
\item{xs}{An integer vector to count populations}
}
\value{
The populations of elements in the vector
}
\description{
Count 1's in each integer element
}
