% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popcount_cpp_raw}
\alias{popcount_cpp_raw}
\title{Count 1's in each raw element}
\usage{
popcount_cpp_raw(xs)
}
\arguments{
\item{xs}{A raw vector to count populations}
}
\value{
The populations of elements in the vector
}
\description{
Count 1's in each raw element
}
